requirements.txt
Django==4.2.2
python-dotenv==1.0
djangorestframework==3.14
drf-spectacular==0.16